+++ debian/changelog
@@ -1,4 +1,5 @@
   [Michele Simionato]
+  * Moved commonlib.parallel into baselib
   * Merged the experimental calculator ebrisk into event_based_risk and
     used correctly the random_seed for generating the GMFs (not the master_seed)
   * Added a flag `ignore_covs` to ignore the coefficients of variation
