+++ doc/sphinx/openquake.commonlib.rst
@@ -57,10 +57,10 @@ openquake.commonlib.oqvalidation module
     :undoc-members:
     :show-inheritance:
 
-openquake.commonlib.parallel module
+openquake.baselib.parallel module
 -----------------------------------
 
-.. automodule:: openquake.commonlib.parallel
+.. automodule:: openquake.baselib.parallel
     :members:
     :undoc-members:
     :show-inheritance:
