+++ openquake/calculators/base.py
@@ -37,7 +37,7 @@ from openquake.hazardlib.calc.filters import SourceFilter
 from openquake.risklib import riskinput, __version__ as engine_version
 from openquake.commonlib import readinput, datastore, source, calc
 from openquake.commonlib.oqvalidation import OqParam
-from openquake.commonlib.parallel import starmap, executor, wakeup_pool
+from openquake.baselib.parallel import starmap, executor, wakeup_pool
 from openquake.baselib.python3compat import with_metaclass
 from openquake.commonlib.export import export as exp
