+++ openquake/calculators/disaggregation.py
@@ -28,7 +28,8 @@ from openquake.baselib import hdf5
 from openquake.baselib.general import split_in_blocks
 from openquake.hazardlib.calc import disagg
 from openquake.hazardlib.calc.filters import SourceFilter
-from openquake.commonlib import parallel, sourceconverter
+from openquake.baselib import parallel
+from openquake.commonlib import sourceconverter
 from openquake.calculators import base, classical
 
 DISAGG_RES_FMT = 'disagg/poe-%(poe)s-rlz-%(rlz)s-%(imt)s-%(lon)s-%(lat)s'
