+++ openquake/commands/tests/independence_test.py
@@ -39,8 +39,7 @@ class IndependenceTestCase(unittest.TestCase):
     def test_commonlib(self):
         assert_independent('openquake.commonlib', 'openquake.calculators')
         assert_independent('openquake.commonlib.node', 'openquake.hazardlib')
-        # NB: ugly: parallel uses GroundShakingIntensityModel
-        # assert_independent('openquake.commonlib.parallel', 'openquake.hazardlib')
+        assert_independent('openquake.baselib.parallel', 'openquake.hazardlib')
 
     def test_engine(self):
         assert_independent('openquake.engine', 'openquake.server')
