+++ openquake/commonlib/oqvalidation.py
@@ -20,11 +20,12 @@ import os
 import logging
 import numpy
 
+from openquake.baselib import parallel
 from openquake.baselib.general import DictArray
 from openquake.hazardlib.imt import from_string
 from openquake.hazardlib import correlation
 from openquake.risklib import valid
-from openquake.commonlib import parallel, logictree
+from openquake.commonlib import logictree
 from openquake.commonlib.riskmodels import get_risk_files
 
 GROUND_MOTION_CORRELATION_MODELS = ['JB2009']
