+++ openquake/server/dbserver.py
@@ -27,7 +27,7 @@ from multiprocessing.connection import Listener
 from concurrent.futures import ProcessPoolExecutor
 
 from openquake.baselib import sap
-from openquake.commonlib.parallel import safely_call
+from openquake.baselib.parallel import safely_call
 from openquake.risklib import valid
 from openquake.commonlib import config
 from openquake.server.db import actions
